Main Menu + Hookshot changes.
Main Menu speaks for itself. It should have an environment, but that's not my choice, and Daniel was ok with me changing the camera perspective per option.
The hookshot was slightly changed (I don't agree with it but whatever). You can't shoot under yourself anymore.

Y m ai wryting in engrish?